Move getPersonaFromConfig into persona folder
