+++ packages/sanes-chrome-extension/src/logic/persona/index.ts
@@ -1,1 +1,2 @@
+export * from './fromconfig';
 export * from './persona';
